Move database cleanup routine into DatabaseCleaner type
